ボス部屋に入った時の処理を修正中…
