+++ Assets/Prefabs/Enemy/Boss_CutSceneTimeline.playable
@@ -81,6 +81,7 @@ MonoBehaviour:
   m_FixedDuration: 0
   m_EditorSettings:
     m_Framerate: 60
+    m_ScenePreview: 1
   m_DurationMode: 0
   m_MarkerTrack: {fileID: -538786167319595129}
 --- !u!114 &1287269922826108375
