--- Assets/Scripts/Runtime/NPC/Boss/BossRoom.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 93d35523933cf0342ae6eb86d32979c4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
